Fix packer vsphere-iso builder.
When one's vsphere data center(s) are contained within a data center
folder, vsphere-iso is broken.   There is no way to specificy a "folder"
in the packer template to allow builds to complete.  See
https://github.com/jetbrains-infra/packer-builder-vsphere/issues/285 for
more details.

This change and removal of the folder template key, or using a folder
value that includes "vm/..." allows successful builds.  It is not clear
why vm/ must be in this key and is also in the code.

DCO 1.1 Signed-off-by: David Ohlemacher <dohlemacher@infinidat.com>